Do it again

+++ src/MSL/string.h
@@ -9,7 +9,7 @@ int __StringRead(void* isc, int ch, int Action);
 char* strchr(const char* str, int chr);
 int strncmp(const char* str1, const char* str2, unsigned long n);
 int strcmp(const char* str1, const char* str2);
-char*(strcpy) (char* __restrict dst, const char* __restrict src);
-unsigned long(strlen)(const char* s);
+char* strcpy(char* __restrict dst, const char* __restrict src);
+size_t strlen(const char* s);
 
 #endif
